+++ src/Microsoft.OpenApi.Readers/OpenApiTextReaderReader.cs
@@ -1,16 +1,11 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license. 
 
-using System;
 using System.IO;
 using System.Linq;
-using Microsoft.OpenApi.Exceptions;
-using Microsoft.OpenApi.Extensions;
 using Microsoft.OpenApi.Interfaces;
 using Microsoft.OpenApi.Models;
 using Microsoft.OpenApi.Readers.Interface;
-using Microsoft.OpenApi.Readers.Services;
-using Microsoft.OpenApi.Services;
 using SharpYaml;
 using SharpYaml.Serialization;
 
@@ -40,9 +35,7 @@ namespace Microsoft.OpenApi.Readers
         /// <returns>Instance of newly created OpenApiDocument</returns>
         public OpenApiDocument Read(TextReader input, out OpenApiDiagnostic diagnostic)
         {
-            ParsingContext context;
             YamlDocument yamlDocument;
-            diagnostic = new OpenApiDiagnostic();
 
             // Parse the YAML/JSON
             try
@@ -51,57 +44,12 @@ namespace Microsoft.OpenApi.Readers
             }
             catch (YamlException ex)
             {
+                diagnostic = new OpenApiDiagnostic();
                 diagnostic.Errors.Add(new OpenApiError($"#char={ex.Start.Line}", ex.Message));
                 return new OpenApiDocument();
             }
 
-            context = new ParsingContext
-            {
-                ExtensionParsers = _settings.ExtensionParsers,
-                BaseUrl = _settings.BaseUrl
-            };
-
-            OpenApiDocument document = null;
-
-            try
-            {
-                // Parse the OpenAPI Document
-                document = context.Parse(yamlDocument, diagnostic);
-
-                // Resolve References if requested
-                switch (_settings.ReferenceResolution)
-                {
-                    case ReferenceResolutionSetting.ResolveAllReferences:
-                        throw new ArgumentException(Properties.SRResource.CannotResolveRemoteReferencesSynchronously);
-                    case ReferenceResolutionSetting.ResolveLocalReferences:
-                        var resolver = new OpenApiReferenceResolver(document);
-                        var walker = new OpenApiWalker(resolver);
-                        walker.Walk(document);
-                        foreach (var item in resolver.Errors)
-                        {
-                            diagnostic.Errors.Add(item);
-                        }
-                        break;
-                    case ReferenceResolutionSetting.DoNotResolveReferences:
-                        break;
-                }
-            }
-            catch (OpenApiException ex)
-            {
-                diagnostic.Errors.Add(new OpenApiError(ex));
-            }
-
-            // Validate the document
-            if (_settings.RuleSet != null && _settings.RuleSet.Rules.Count > 0)
-            {
-                var errors = document.Validate(_settings.RuleSet);
-                foreach (var item in errors)
-                {
-                    diagnostic.Errors.Add(item);
-                }
-            }
-
-            return document;
+            return new OpenApiYamlDocumentReader(this._settings).Read(yamlDocument, out diagnostic);
         }
         /// <summary>
         /// Reads the stream input and parses the fragment of an OpenAPI description into an Open API Element.
@@ -112,9 +60,7 @@ namespace Microsoft.OpenApi.Readers
         /// <returns>Instance of newly created OpenApiDocument</returns>
         public T ReadFragment<T>(TextReader input, OpenApiSpecVersion version, out OpenApiDiagnostic diagnostic) where T : IOpenApiElement
         {
-            ParsingContext context;
             YamlDocument yamlDocument;
-            diagnostic = new OpenApiDiagnostic();
 
             // Parse the YAML/JSON
             try
@@ -123,38 +69,12 @@ namespace Microsoft.OpenApi.Readers
             }
             catch (YamlException ex)
             {
+                diagnostic = new OpenApiDiagnostic();
                 diagnostic.Errors.Add(new OpenApiError($"#line={ex.Start.Line}", ex.Message));
                 return default(T);
             }
 
-            context = new ParsingContext
-            {
-                ExtensionParsers = _settings.ExtensionParsers
-            };
-
-            IOpenApiElement element = null;
-
-            try
-            {
-                // Parse the OpenAPI element
-                element = context.ParseFragment<T>(yamlDocument, version, diagnostic);
-            }
-            catch (OpenApiException ex)
-            {
-                diagnostic.Errors.Add(new OpenApiError(ex));
-            }
-
-            // Validate the element
-            if (_settings.RuleSet != null && _settings.RuleSet.Rules.Count > 0)
-            {
-                var errors = element.Validate(_settings.RuleSet);
-                foreach (var item in errors)
-                {
-                    diagnostic.Errors.Add(item);
-                }
-            }
-
-            return (T)element;
+            return new OpenApiYamlDocumentReader(this._settings).ReadFragment<T>(yamlDocument, version, out diagnostic);
         }
 
         /// <summary>
